Module Tezos_mockup.Local_services

module Directory : sig ... end

type Tezos_base__TzPervasives.error +=

| Injection_not_possible

type Tezos_base__TzPervasives.error +=

| Cannot_parse_op

module type MENV = sig ... end

module Make (E : MENV) : sig ... end

val build_shell_directory : string -> Tezos_mockup_registration.Registration.M.mockup_environment -> Tezos_crypto.Chain_id.t -> Tezos_protocol_environment.rpc_context -> bool -> (Tezos_protocol_environment.rpc_context -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t) -> unit Directory.directory

val build_directory : string -> bool -> Tezos_mockup_registration.Registration.M.mockup_environment -> Tezos_crypto.Chain_id.t -> Tezos_protocol_environment.rpc_context -> unit Tezos_rpc.RPC_directory.t

The directory of RPCs that the mockup client honors. Parameters are:

mem_only specifies whether the mockup uses a persistent state. mockup_env is the implementation provided by the protocol. chain_id is the only chain that the mockup honors. rpc_context is data used when honoring an RPC.