Module type Local_services.MENV

include Tezos_mockup_registration.Registration_intf.MOCKUP

type parameters

type protocol_constants

val parameters_encoding : parameters Tezos_base__TzPervasives.Data_encoding.t

val default_parameters : parameters

val protocol_constants_encoding : protocol_constants Tezos_base__TzPervasives.Data_encoding.t

val default_protocol_constants : Tezos_client_base.Client_context.full -> (protocol_constants, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

The content equivalent to the default value of --protocol-constants

val default_bootstrap_accounts : Tezos_client_base.Client_context.full -> (string, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

The content equivalent to the default value of --bootstrap-accounts

val protocol_hash : Tezos_crypto.Protocol_hash.t

module Protocol : Tezos_mockup_registration.Registration_intf.PROTOCOL

module Block_services : sig ... end

val directory : Tezos_protocol_environment.rpc_context Tezos_rpc.RPC_directory.t

val init : cctxt:Tezos_client_base.Client_context.full -> parameters:parameters -> constants_overrides_json:Tezos_base__TzPervasives.Data_encoding.json option -> bootstrap_accounts_json:Tezos_base__TzPervasives.Data_encoding.json option -> (Tezos_mockup_registration.Registration_intf.T.mockup_context, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val migrate : Tezos_mockup_registration.Registration_intf.T.mockup_context -> (Tezos_mockup_registration.Registration_intf.T.mockup_context, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val chain_id : Tezos_crypto.Chain_id.t

val rpc_context : Tezos_protocol_environment.rpc_context

val base_dir : string